requirements.txt
asgiref     
cycler              
decorator             
Django                  
h5py                 
imageio              
Keras>=2.3.1                 
Keras-Applications    
Keras-Preprocessing 
tensorflow>=2.0.0
tensorflow-estimator==2.0.0
kiwisolver      
matplotlib      
mysqlclient     
networkx          
numpy          
opencv-python
Pillow   
pip  
pyparsing   
python-dateutil
pytz  
PyWavelets
PyYAML     
scikit-image  
scipy
setuptools  
six  
sqlparse
uWSGI  
wheel
